deps.sh: Correctly detect the apt-utility lacking install-support
In earlier versions of Ubuntu (like 12.04) and other Debian-based distributions, the APT-utility exists without support for the install-command.
This causes execution of deps.sh to fail since c8cc356 when running on these older versions of apt. This commit solves this issue by attempting
a simulated install of dpkg.

+++ deps.sh
@@ -187,10 +187,18 @@ then
         PKG_PYTHON3_CANDIDATES="$PYTHON3_APT_CANDIDATES"
     fi
 else
-    PKGCMD="$APTCMD"
-    PKGCMD_OPTS="install -y"
-    PKG_CANDIDATES="$APT_CANDIDATES"
-    PKG_PYTHON3_CANDIDATES="$PYTHON3_APT_CANDIDATES"
+    if "$APTCMD" install -s -y dpkg > /dev/null
+    then
+        PKGCMD="$APTCMD"
+        PKGCMD_OPTS="install -y"
+        PKG_CANDIDATES="$APT_CANDIDATES"
+        PKG_PYTHON3_CANDIDATES="$PYTHON3_APT_CANDIDATES"
+    else
+        PKGCMD="$APTGETCMD"
+        PKGCMD_OPTS="install -y"
+        PKG_CANDIDATES="$APT_CANDIDATES"
+        PKG_PYTHON3_CANDIDATES="$PYTHON3_APT_CANDIDATES"
+    fi
 fi
 
 if [ "$NEEDED_UTILS" != "" ]
